Moved clubmembership ops to service function
